requirements.txt
python-dotenv
pytest
pytest-mock
python-telegram-bot[ext]
pytest-asyncio
pypdf
requests 